Major change made to AlignmentModel.Cache
When updating the cache, if the number of reads being added to the
interval tree is more than 500,000 the free memory available is checked.
If the memory used is more than 80% and thus, the free memory is less
than 20% of the total memory available, the cache is aborted and we
return the iterator over the entire region without using the cache. The
problem regions are stored in the cache class object so that if the
program tries to get reads over this region again, it will know that it
must not use